requirements.txt
streamlit
tensorflow
numpy
Pillow
keras
requesta